refactor: MessageFixtures -> MessageFixture로 이름 변경
