+++ backend/src/test/java/com/pickpick/slackevent/application/channel/ChannelDeletedServiceTest.java
@@ -2,8 +2,8 @@ package com.pickpick.slackevent.application.channel;
 
 import static com.pickpick.fixture.ChannelFixture.NOTICE;
 import static com.pickpick.fixture.MemberFixture.HOPE;
-import static com.pickpick.fixture.MessageFixtures.PLAIN_20220712_18_00_00;
-import static com.pickpick.fixture.MessageFixtures.PLAIN_20220715_14_00_00;
+import static com.pickpick.fixture.MessageFixture.PLAIN_20220712_18_00_00;
+import static com.pickpick.fixture.MessageFixture.PLAIN_20220715_14_00_00;
 import static com.pickpick.fixture.WorkspaceFixture.JUPJUP;
 import static com.pickpick.support.JsonUtils.toJson;
 import static org.assertj.core.api.Assertions.assertThat;
